requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
colorama==0.4.6
execnet==2.0.2
Faker==23.1.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pip==24.0
pluggy==1.4.0
pycparser==2.21
PySocks==1.7.1
pytest-xdist==3.5.0
pytest==8.0.0
python-dateutil==2.8.2
selenium==4.17.2
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio-websocket==0.11.1
trio==0.24.0
typing-extensions==4.9.0
urllib3==2.2.0
wheel==0.42.0
wsproto==1.2.0
